microservices/formulas/pyproject.toml
[tool.poetry]
name = "formulas"
version = "0.1.0"
description = ""
authors = ["Kevin <zhijunwang2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-cors-headers = "^3.13.0"
pika = "^1.3.1"


[tool.poetry.group.dev.dependencies]

aiohttp = "3.8.3"
aiosignal = "1.3.1"
alembic = "1.9.2"
asgiref = "3.6.0"
async-timeout = "4.0.2"
asyncio = "3.4.3"
attrs = "22.2.0"
cached-property = "1.5.2"
charset-normalizer = "2.1.1"
click = "8.1.3"
colorama = "0.4.6"
dependency-injector = "4.41.0"
distlib = "0.3.6"
django = "4.1.6"
djangorestframework = "3.14.0"
filelock = "3.9.0"
flask = "2.2.2"
flask-cors = "3.0.10"
flask-migrate = "4.0.2"
flask-sqlalchemy = "3.0.2"
frozenlist = "1.3.3"
greenlet = "2.0.1"
h11 = "0.14.0"
idna = "3.4"
inflection = "0.5.1"
itsdangerous = "2.1.2"
jinja2 = "3.0.3"
mako = "1.2.4"
markupsafe = "2.1.1"
multidict = "6.0.4"
mypy-extensions = "0.4.3"
platformdirs = "2.6.2"
psycopg2-binary = "2.9.5"
python-dotenv = "0.21.1"
pytz = "2022.7.1"
six = "1.16.0"
sqlalchemy = "1.4.46"
sqlalchemy-orm = "1.2.3"
sqlparse = "0.4.3"
typing-inspect = "0.8.0"
typing-extensions = "4.4.0"
tzdata = "2022.7"
uvicorn = "0.20.0"
virtualenv = "20.17.1"
werkzeug = "2.2.2"
yarl = "1.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
